+++ .settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<faceted-project>
+  <fixed facet="wst.jsdt.web"/>
+  <installed facet="java" version="1.7"/>
+  <installed facet="jst.web" version="3.1"/>
+  <installed facet="jst.jsf" version="2.2"/>
+  <installed facet="jst.jaxrs" version="2.0"/>
+  <installed facet="wst.jsdt.web" version="1.0"/>
+</faceted-project>
