+++ pom.xml
@@ -46,6 +46,43 @@
             <artifactId>java-jwt</artifactId>
             <version>3.2.0</version>
         </dependency>
+          <dependency>
+		    <groupId>org.eclipse.jetty</groupId>
+		    <artifactId>jetty-servlet</artifactId>
+		    <version>9.2.3.v20140905</version>
+		</dependency>
+		<dependency>
+		     <groupId>org.glassfish.jersey.core</groupId>
+		     <artifactId>jersey-server</artifactId>
+		     <version>2.7</version>
+		 </dependency>
+		
+		<dependency>
+		     <groupId>org.glassfish.jersey.containers</groupId>
+		     <artifactId>jersey-container-servlet-core</artifactId>
+		     <version>2.7</version>
+		 </dependency>
+        <dependency>
+		    <groupId>org.glassfish.jersey.containers</groupId>
+		    <artifactId>jersey-container-jetty-http</artifactId>
+		    <version>2.25.1</version>
+		</dependency>
+		<!-- LOGGING -->
+		<dependency>
+				<groupId>log4j</groupId>
+				<artifactId>log4j</artifactId>
+				<version>1.2.17</version>
+			</dependency>
+			 <dependency>
+	       <groupId>org.slf4j</groupId>
+	       <artifactId>slf4j-api</artifactId>
+	       <version>1.7.5</version>
+	   </dependency>
+	   <dependency>
+	       <groupId>org.slf4j</groupId>
+	       <artifactId>slf4j-log4j12</artifactId>
+	       <version>1.7.5</version>
+	   </dependency>
     </dependencies>
 
     <build>
@@ -74,7 +111,7 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-dependency-plugin</artifactId>
                 <version>2.6</version>
-                <executions>
+                <!-- <executions>
                     <execution>
                         <phase>validate</phase>
                         <goals>
@@ -93,7 +130,7 @@
                             </artifactItems>
                         </configuration>
                     </execution>
-                </executions>
+                </executions> -->
             </plugin>
         </plugins>
     </build>
